# src/main/resources/apidocs.json
{
    "swagger": "2.0",
    "info": {
        "description": "FiestaReasonerEngine API documentation",
        "version": "0.0.1",
        "title": "FiestaReasonerEngine API",
        "contact": {},
        "license": {}
    },
    "host": "platform-dev.fiesta-iot.eu",
    "basePath": "/reasoner-engine",
...
    "paths": {
        "/api/executions": {
            "get": {
                "tags": [
                    "execution-resource"
                ],
                "summary": "getAllExecutions",
                "operationId": "getAllExecutionsUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number of the requested page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of a page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Execution"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "execution-resource"
                ],
                "summary": "createExecution",
                "operationId": "createExecutionUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "execution",
                        "description": "execution",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ExecuteRuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Execution"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error!"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/executions/{id}": {
            "get": {
                "tags": [
                    "execution-resource"
                ],
                "summary": "getExecution",
                "operationId": "getExecutionUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Execution"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ExecuteRuleRequest": {
            "type": "object",
            "required": [
                "executeType",
                "registerRuleId"
            ],
            "properties": {
                "ended": {
                    "type": "string",
                    "format": "date-time"
                },
                "executeType": {
                    "type": "integer",
                    "format": "int32"
                },
                "registerRuleId": {
                    "type": "integer",
                    "format": "int64"
                },
                "started": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "Execution": {
            "type": "object",
            "required": [
                "registerRule"
            ],
            "properties": {
                "created": {
                    "type": "string",
                    "format": "date-time"
                },
                "ended": {
                    "type": "string",
                    "format": "date-time"
                },
                "fullData": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "infferedData": {
                    "type": "string"
                },
                "originalData": {
                    "type": "string"
                },
                "registerRule": {
                    "$ref": "#/definitions/RegisterRule"
                },
                "ruleContent": {
                    "type": "string"
                },
                "sensor": {
                    "type": "string"
                },
                "started": {
                    "type": "string",
                    "format": "date-time"
                },
                "status": {
                    "type": "boolean"
                },
                "type": {
                    "type": "integer",
                    "format": "int32"
                },
                "updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "Reasoning": {
            "type": "object",
            "required": [
                "content",
                "name",
                "sensor"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "created": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string"
                },
                "hashedSensor": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "latitude": {
                    "type": "number",
                    "format": "float"
                },
                "longitude": {
                    "type": "number",
                    "format": "float"
                },
                "name": {
                    "type": "string"
                },
                "quantityKind": {
                    "type": "string"
                },
                "ruleType": {
                    "type": "integer",
                    "format": "int32"
                },
                "sensor": {
                    "type": "string"
                },
                "sensorEndp": {
                    "type": "string"
                },
                "sensorMeta": {
                    "type": "string"
                },
                "sensorSampleData": {
                    "type": "string"
                },
                "unitOfMeasurement": {
                    "type": "string"
                },
                "updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "RegisterRule": {
            "type": "object",
            "required": [
                "data",
                "name",
                "reasoning",
                "ruleContent",
                "sensor"
            ],
            "properties": {
                "created": {
                    "type": "string",
                    "format": "date-time"
                },
                "data": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fullData": {
                    "type": "string"
                },
                "hashedSensor": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "inferredData": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number",
                    "format": "float"
                },
                "longitude": {
                    "type": "number",
                    "format": "float"
                },
                "name": {
                    "type": "string"
                },
                "quantityKind": {
                    "type": "string"
                },
                "reasoning": {
                    "$ref": "#/definitions/Reasoning"
                },
                "ruleContent": {
                    "type": "string"
                },
                "sensor": {
                    "type": "string"
                },
                "sensorEndp": {
                    "type": "string"
                },
                "sensorMeta": {
                    "type": "string"
                },
                "unitOfMeasurement": {
                    "type": "string"
                },
                "updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    }
}